+++ packages/chain-source/CHANGELOG.md
@@ -6,7 +6,7 @@ this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.6.0] — 2026-05-05
 
 ### Added